// Frontend/src/api/models/getProgramsWithFilterQuery.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * AuditSupport
 * OpenAPI spec version: v1
 */
import type { PaginationDTO } from './paginationDTO';
import type { ProgramStatuses } from './programStatuses';

export interface GetProgramsWithFilterQuery {
  yearId?: string;
  /** @nullable */
  teachers?: string[] | null;
  pagination?: PaginationDTO;
  /** @nullable */
  statuses?: ProgramStatuses[] | null;
  /** @nullable */
  directions?: string[] | null;
}