// Frontend/src/api/models/updateUserRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * AuditSupport
 * OpenAPI spec version: v1
 */
import type { Roles } from './roles';

export interface UpdateUserRequest {
  userId?: string;
  /** @nullable */
  surname?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  patronymic?: string | null;
  /** @nullable */
  email?: string | null;
  role?: Roles;
  isActive?: boolean;
}